// apps/desktop/src/generated/GitFileDiffDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GitDiffHunkDto } from "./GitDiffHunkDto";

export type GitFileDiffDto = {
  repositoryId: string;
  path: string;
  originalPath: string | null;
  additions: number;
  deletions: number;
  binary: boolean;
  truncated: boolean;
  hunks: Array<GitDiffHunkDto>;
};