// apps/desktop/src/generated/GitSnapshotDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GitRepositoryStatusDto } from "./GitRepositoryStatusDto";

export type GitSnapshotDto = {
  spaceId: string;
  rootId: string;
  scanGeneration: number;
  repositories: Array<GitRepositoryStatusDto>;
};